// src/file-service-api/v1/model/userRegistrationRequest.ts
/**
 * Vorba File Service (Development)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserRegistrationRequest { 
    /**
     * User email address
     */
    email: string;
    /**
     * User password
     */
    password: string;
    /**
     * User full name
     */
    name: string;
    /**
     * Username
     */
    username: string;
}